Move functions back to FetchBlogListWPOrg
So it will be easier to merge with Danilo's updates to the original code.
